backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pandas==2.2.1
python-dotenv==1.0.1
pytest-django==4.8.0
pytest==8.1.1
